Rename session database to control database.
This is 100% find and replace operations. The gist is changing "session" to "control" almost every place it shows up. This is because once we have finished the initial handshake, the control database contains data that is global to all of a user's sessions. So I think it will be long-term less confusing to call it the control database in most places.